Fix race condtion with closing
When we close while waiting for the next periodic re-roll or ScanNow we
previously would allow the loop to execute one last time, start a rescan,
and then exit the Start() function when we got a <-w.close.  However,
when this happens and there's a changed files there's a 50% for each file
we'll choose to process the <-evt file event first before the <-w.close!
This is a huge problem if it's no longer safe to do so.

This change simply exits the Start function (safely, closing all the
channels as before) as soon as we receive a <-w.close